updated files

+++ src/Controller/StudentsController.php
@@ -2,6 +2,8 @@
 namespace App\Controller;
 
 use App\Controller\AppController;
+use App\Controller\Traits\SearchParameterTrait;
+use Cake\Datasource\Exception\RecordNotFoundException;
 
 /**
  * Students Controller
@@ -11,6 +13,7 @@ use App\Controller\AppController;
 class StudentsController extends AppController
 {
 
+    use SearchParameterTrait;
     /**
      * Index method
      *
@@ -18,12 +21,35 @@ class StudentsController extends AppController
      */
     public function index()
     {
-        $this->paginate = [
-            'contain' => ['Sessions', 'Classes', 'ClassDemacations']
-        ];
+        if ( empty($this->request->query['class_id'])) {
+            $this->paginate = [
+                'limit' => 50,
+                'contain' => ['Sessions', 'Classes'],
+                'conditions' => [
+                    'Students.status'   => 1,
+                    'Students.graduated'   => 0
+                ],
+                // Place the result in ascending order according to the class.
+                'order' => [
+                    'class_id' => 'asc'
+                ]
+            ];
+        }
+        else {
+            $this->paginate = [
+                'limit' => 50,
+                'contain' => ['Sessions', 'Classes'],
+                'conditions' => [
+                    'Students.status'   => 1,
+                    'Students.graduated'   => 0,
+                    'Students.class_id' => $this->_getDefaultValue($this->request->query['class_id'],1)
+                ]
+            ];
+        }
         $students = $this->paginate($this->Students);
-
-        $this->set(compact('students'));
+        $sessions = $this->Students->Sessions->find('list',['limit' => 200]);
+        $classes = $this->Students->Classes->find('list',['limit' => 200]);
+        $this->set(compact('students','sessions','classes'));
         $this->set('_serialize', ['students']);
     }
 
@@ -36,12 +62,23 @@ class StudentsController extends AppController
      */
     public function view($id = null)
     {
-        $student = $this->Students->get($id, [
-            'contain' => ['Sessions', 'Classes', 'ClassDemacations']
-        ]);
+        try {
+            /**
+             * if no @var $id was specified redirect to the index action
+             */
+            if (empty($id)) {
+                $this->redirect(['action' => 'index']);
+            }
+            $student = $this->Students->get($id, [
+                'contain' => ['Sessions', 'Classes','SessionAdmitted']
+            ]);
+            $this->set('student', $student);
+            $this->set('_serialize', ['student']);
 
-        $this->set('student', $student);
-        $this->set('_serialize', ['student']);
+        } catch ( RecordNotFoundException $e ) {
+            $this->render('/Element/Error/recordnotfound');
+        }
+
     }
 
     /**
@@ -57,15 +94,15 @@ class StudentsController extends AppController
             if ($this->Students->save($student)) {
                 $this->Flash->success(__('The student has been saved.'));
 
-                return $this->redirect(['action' => 'index']);
+                return $this->redirect(['action' => 'add']);
             } else {
                 $this->Flash->error(__('The student could not be saved. Please, try again.'));
             }
         }
         $sessions = $this->Students->Sessions->find('list', ['limit' => 200]);
         $classes = $this->Students->Classes->find('list', ['limit' => 200]);
-        $classDemacations = $this->Students->ClassDemacations->find('list', ['limit' => 200]);
-        $this->set(compact('student', 'sessions', 'classes', 'classDemacations'));
+        $classDemarcations = $this->Students->ClassDemarcations->find('list', ['limit' => 200]);
+        $this->set(compact('student', 'sessions', 'classes', 'classDemarcations'));
         $this->set('_serialize', ['student']);
     }
 
@@ -78,23 +115,34 @@ class StudentsController extends AppController
      */
     public function edit($id = null)
     {
-        $student = $this->Students->get($id, [
-            'contain' => []
-        ]);
+        try {
+            /**
+             * if no @var $id was specified redirect to the index action
+             */
+            if (empty($id)) {
+                $this->redirect(['action' => 'index']);
+            }
+            $student = $this->Students->get($id, [
+                'contain' => []
+            ]);
+        } catch ( RecordNotFoundException $e ) {
+            $this->Flash->error('Record Not found');
+        }
+
         if ($this->request->is(['patch', 'post', 'put'])) {
             $student = $this->Students->patchEntity($student, $this->request->data);
-            if ($this->Students->save($student)) {
+            if ( $this->Students->save($student) ) {
                 $this->Flash->success(__('The student has been saved.'));
 
-                return $this->redirect(['action' => 'index']);
+                //return $this->redirect(['action' => 'index']);
             } else {
                 $this->Flash->error(__('The student could not be saved. Please, try again.'));
             }
         }
         $sessions = $this->Students->Sessions->find('list', ['limit' => 200]);
         $classes = $this->Students->Classes->find('list', ['limit' => 200]);
-        $classDemacations = $this->Students->ClassDemacations->find('list', ['limit' => 200]);
-        $this->set(compact('student', 'sessions', 'classes', 'classDemacations'));
+        $classDemarcations = $this->Students->ClassDemarcations->find('list', ['limit' => 200]);
+        $this->set(compact('student', 'sessions', 'classes', 'classDemarcations'));
         $this->set('_serialize', ['student']);
     }
 
@@ -107,14 +155,37 @@ class StudentsController extends AppController
      */
     public function delete($id = null)
     {
-        $this->request->allowMethod(['post', 'delete']);
-        $student = $this->Students->get($id);
-        if ($this->Students->delete($student)) {
-            $this->Flash->success(__('The student has been deleted.'));
-        } else {
-            $this->Flash->error(__('The student could not be deleted. Please, try again.'));
-        }
+        if (empty($id)) {
+            $this->redirect(['action' => 'index']);
+            $this->Flash->error('Please Specify the student Id was not specified. Try again');
 
-        return $this->redirect(['action' => 'index']);
+        } else {
+            $this->request->allowMethod(['post', 'delete']);
+            $student = $this->Students->get($id);
+            if ($this->Students->delete($student)) {
+                $this->Flash->success(__('The student has been deleted.'));
+            } else {
+                $this->Flash->error(__('The student could not be deleted. Please, try again.'));
+            }
+
+            return $this->redirect(['action' => 'index']);
+        }
+    }
+
+    /**
+     *  The function take now argument .
+     * It returns the graduated students
+     */
+    public function graduatedStudents()
+    {
+        // get the graduated Students .
+        $graduatedStudents = $this->Students->find('GraduatedStudents');
+        $this->set(compact('graduatedStudents'));
+    }
+
+    public function unActiveStudents()
+    {
+        $unActiveStudents = $this->Students->findUnActiveStudents();
+        $this->set(compact('unActiveStudents'));
     }
 }
